+++ inc/ti/archive.h
@@ -28,4 +28,5 @@ int ti_archive_to_disk(void);
 uint64_t ti_archive_get_first_change_id(void);
 ti_cpkg_t * ti_archive_get_change(uint64_t change_id);
 
+
 #endif /* TI_ARCHIVE_H_ */