+++ src/ti/task.c
@@ -2390,6 +2390,45 @@ fail_data:
     return -1;
 }
 
+int ti_task_add_fill(ti_task_t * task, ti_raw_t * key, ti_val_t * val)
+{
+    size_t alloc = 512;  /* often used for small values */
+    ti_data_t * data;
+    msgpack_packer pk;
+    msgpack_sbuffer buffer;
+
+    if (mp_sbuffer_alloc_init(&buffer, alloc, sizeof(ti_data_t)))
+        return -1;
+    msgpack_packer_init(&pk, &buffer, msgpack_sbuffer_write);
+
+    msgpack_pack_array(&pk, 2);
+
+    msgpack_pack_uint8(&pk, TI_TASK_FILL);
+    msgpack_pack_map(&pk, 1);
+
+    mp_pack_strn(&pk, key->data, key->n);
+
+    if (ti_val_gen_ids(val) || ti_val_to_store_pk(val, &pk))
+        goto fail_pack;
+
+    data = (ti_data_t *) buffer.data;
+    ti_data_init(data, buffer.size);
+
+    if (vec_push(&task->list, data))
+        goto fail_data;
+
+    task__upd_approx_sz(task, data);
+    return 0;
+
+fail_data:
+    free(data);
+    return -1;
+
+fail_pack:
+    msgpack_sbuffer_destroy(&buffer);
+    return -1;
+}
+
 int ti_task_add_splice(
         ti_task_t * task,
         ti_raw_t * key,