+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.4.2
+
+* Added `fill(..)` function, issue #320.
+* Allow away mode (blocking) on a single node, issue #321.
+* Increase `range(..)` limitation, issue #322.
+
 # v1.4.1
 
 * Return same instance on type match, issue #315.